Bind to specific port.
We encountered connection problems with the Java API. First suspicion was the fact that multicast packages were sent via IPv6 so we switched that off system wide. Next error message involved the client connecting to some 127.0.1.1 address, so we added that to the lo interface and let the server bind to that address. Remains one issue: When connecting to the server now we get an exception on the server side involving some 192.168.*.* address in case connection fails.