Second approach to using ES for input.
Uses the ES termvectors query to retrieve the pre-analysed/tokenised vectors of
ES as input to the Mahout feature vector encoder. With the current encoding all
term vectors need to be transferred to a central node that trains the logistic
regression model. Wondering whether this could somehow be done in parallel on
each node - combining the models afterwards...